Record whether the BSDF PDF is proportional in BSDFSample
(And eliminate the SampledPDFIsProportional() method from the BxDF
interface.)

This fixes a previously-unknown(*) bug that was causing noise in some
scenes. The issue is that the LayeredBxDF actually doesn't always return a
proportional PDF: if the path scatters at the interface and doesn't enter
the layers, the BSDFSample is a regular one. The PDF is only proportional
if the path starts bouncing around in there...

(*) This could be the issue noted in #14--not yet confirmed...