Re-plumb texture creation to create both SpectrumType::Reflectance and General variants
When a texture is first created, we don't know how it will be
used. Therefore, we create both forms of it and then choose the appropriate
one based on what the Material requests. This eliminates a bunch of cases
where we hard-coded creating ones for Reflectance since we didn't know any
better.