Fix various bugs in the type of RGB Spectrum being used.
In many cases we ended up using a spectrum with the illuminant baked in
when we didn't actually want it.  This caused a blue color shift when equal
RGB values were given to specify scattering and caused the hair BxDF to
come out too red...

(All of these issues were made obvious with the recent
RGBIlluminantSpectrum renaming fixup.)

Fixes, finally, issue #28.