Show OptiX logs on call failure
Some of the OptiX functions fill in a passed-in log variable with
potential information about failures. However that information was only
being printed after the function call and the return value check meaning
it would never be reached in case of failure as abort would be called
before printing the logs.