Rework MediumHandle::SampleTmaj() interface.
Now, it returns a SampledSpectrum representing Tmaj, either from [0,tMax)
or from the last sample position returned via a callback if one or more
samples was taken.  In turn, this value is used to correctly update various
path throughputs and sampling PDFs in the integrators.

This fixes some subtle bugs from mishandling that before. As a bonus, with
this change, volumetrics on the GPU are now about 2x faster, though it's
not obvious why that would be (and I haven't dug in further yet.)