Fix handling of SampledWavelengths in SPPMIntegrator
The old version was all sorts of buggy if one of the BSDFs terminated
secondary wavelengths (including thread races.)

Now we start each photon and camera path from scratch with the
originally-sampled set of wavelengths. Each is allowed to terminate
secondary wavelengths. When photon contributions are added, then we need to
fall back to a single wavelength if either the photon path terminated them
or the visible point is along a camera path that terminated them.