Rework handling of spectral IORs in materials
Now they store a plain Spectrum, even in the case of a constant IOR;
textures are no longer a part of it.

Coated{Diffuse,Conductor} now allow the specification of a spectral
IOR for their top layer; before it could only be a scalar.

This also fixes a seg fault with the dambreak example scene due to a
zero-valued IOR for wavelengths outside of the range of available data.