Improve handling of alpha-masked area lights.
Alpha textures now act on emission the same way they do for visibility:
alpha masked regions of the surface do not emit illumination.

Area lights with a constant alpha of zero are (begrudgingly) special cased
so that it is still possible to specify an invisible area light source.  A
hack in DiffuseAreaLight causes those to be sampled with light sampling
only for direct lighting in integrators so that now they are rendered
correctly. (Before, they would be MIS weighted accounting for the chance of
a BSDF-sampled ray intersecting them even though it was impossible for BSDF
sampled rays to do so...)