GPU: fix multiple significant bugs in OptiXAggregate::createGASForBLPs()
The logic was all kind of borked, with issues including the fact that the
number of primitives in each BLP mesh was reported to be the number of
vertices.  Further, the AABBs weren't being properly initialized, etc.  It
seems to have previously worked for single patch meshes by luck, but was
completely broken for meshes with multiple patches.