VolPathIntegrator: track beta and 1 / path MIS weights.
(Rather than T_maj and path PDFs.)

This is "just" a matter of dividing the old T_maj by the PDF for sampling
the path and then dividing the old path PDFs by that same factor, so the
same mathematically in the end.

However, it a) better aligns the VolPathIntegrator with the other
integrators, which track beta, and b) fixes the issue of T_maj and the PDFs
coming to have very large or very small magnitudes and being succeptible to
over/underflow.  Now, all of the Rescale() stuff is no longer necessary for
them...