Fix bug in RNG seeding in wavefront integrator.
Partially seeding an RNG using just ray direction is risky business with an
orthographic camera. Things go extra bad when the object enclosing the
medium has camera-axis aligned planes in which case doing the rest of the
RNG seeding with tMax to the intersection point rounds out the borkage.

Fixes #175.